requirements.txt
mss
pyzbar
Pillow
pynput
pystray
plyer